+++ index.d.ts
@@ -1,5 +1,5 @@
-export { loguxComponent } from './component/index.js'
-export { createLogux } from './create-logux/index.js'
+export { createStoreCreator } from './store/index.js'
+export { createLogux } from './deprecated/index.js'
 export { LoguxVuex } from './install/index.js'
 export { Channel, loguxMixin, loguxMixinComponent, Subscription } from './mixin/index.js'
-export { createStoreCreator } from './store/index.js'
+export { loguxComponent } from './component/index.js'