+++ install/index.test.js
@@ -1,21 +1,24 @@
 import { mount, createLocalVue } from '@vue/test-utils'
+import { CrossTabClient } from '@logux/client'
 import { TestTime } from '@logux/core'
 
-import { LoguxVuex, createLogux } from '../index.js'
+import { LoguxVuex, createStoreCreator } from '../index.js'
 
 function createComponent (component) {
   let localVue = createLocalVue()
 
   localVue.use(LoguxVuex)
 
-  let Logux = createLogux({
+  let client = new CrossTabClient({
     subprotocol: '1.0.0',
     server: 'wss://localhost:1337',
     userId: '10',
     time: new TestTime()
   })
 
-  let store = new Logux.Store(() => ({ state: { value: 0 } }))
+  let createStore = createStoreCreator(client)
+
+  let store = createStore(() => ({ state: { value: 0 } }))
   let wrapper = mount(component, { localVue, store })
 
   return wrapper